+++ cordex/__init__.py
@@ -2,7 +2,14 @@ import pkg_resources
 
 from . import regions, tables, tutorial
 from .accessor import CordexDataArrayAccessor, CordexDatasetAccessor  # noqa
-from .domain import cordex_domain, create_dataset, domain, domain_info, vertices
+from .domain import (
+    cordex_domain,
+    create_dataset,
+    domain,
+    domain_info,
+    vertices,
+    rewrite_coords,
+)
 from .tables import domains, ecmwf
 from .transform import (
     map_crs,
@@ -41,4 +48,5 @@ __all__ = [
     "domains",
     "ecmwf",
     "cell_area",
+    "rewrite_coords",
 ]